Update status and metadata to CG-DRAFT and fix broken links
The status of the spec was "Unofficial Proposal Draft" and the metadata
associated the spec with the Immersive Web WG, while my understanding is that
the spec is under incubation in the CG (with a plan to adopt it in the WG after
incubation).

This update associates the spec with the CG, bumps the status to CG-DRAFT, and
fixes a few broken links.